Release stable release 'FlipFloppedWhiteSocked'

+++ psef/v1/code.py
@@ -8,6 +8,8 @@ APIs are used to manipulate student submitted code and the related feedback.
 import shutil
 import typing as t
 
+import werkzeug
+import sqlalchemy.sql as sql
 from flask import request, make_response
 from sqlalchemy.orm import make_transient
 
@@ -16,18 +18,14 @@ import psef.files
 import psef.models as models
 import psef.helpers as helpers
 from psef import app, current_user
-from psef.models import db
 from psef.errors import APICodes, APIException
-from psef.models import FileOwner
+from psef.models import FileOwner, db
 from psef.helpers import (
     JSONType, JSONResponse, EmptyResponse, jsonify, ensure_json_dict,
     ensure_keys_in_dict, make_empty_response
 )
 
 from . import api
-
-if t.TYPE_CHECKING:  # pragma: no cover
-    import werkzeug  # NOQA
 
 _HumanFeedback = models.Comment
 _LinterFeedback = t.MutableSequence[t.Tuple[str, models.LinterComment]]
@@ -120,10 +118,9 @@ def remove_comment(id: int, line: int) -> EmptyResponse:
 
 @api.route("/code/<int:file_id>", methods=['GET'])
 @auth.login_required
-def get_code(file_id: int
-             ) -> t.Union['werkzeug.wrappers.Response', JSONResponse[
-                 t.Union[t.Mapping[str, str], models.File, _FeedbackMapping]
-             ]]:
+def get_code(file_id: int) -> t.Union[werkzeug.wrappers.Response, JSONResponse[
+    t.Union[t.Mapping[str, str], models.File, _FeedbackMapping]
+]]:
     """Get data from the :class:`.models.File` with the given id.
 
     .. :quickref: Code; Get code or its metadata.
@@ -132,12 +129,12 @@ def get_code(file_id: int
     argument type in the request different functions are called.
 
     - If ``type == 'metadata'`` the JSON serialized :class:`.models.File` is
-        returned.
+      returned.
     - If ``type == 'file-url'`` or ``type == 'pdf'`` (deprecated) an object
-        with a single key, `name`, with as value the return values of
-        :py:func:`get_file_url`.
+      with a single key, `name`, with as value the return values of
+      :py:func:`.get_file_url`.
     - If ``type == 'feedback'`` or ``type == 'linter-feedback'`` see
-        :py:func:`get_feedback`
+      :py:func:`.code.get_feedback`
     - Otherwise the content of the file is returned as plain text.
 
     :param int file_id: The id of the file
@@ -185,7 +182,7 @@ def get_file_url(file: models.File) -> str:
     return name
 
 
-def get_feedback(file: models.File, linter: bool=False) -> _FeedbackMapping:
+def get_feedback(file: models.File, linter: bool = False) -> _FeedbackMapping:
     """Returns the :class:`.models.Comment` objects attached to the given
     :class:`.models.File` if the user can see them, else returns an empty dict.
 
@@ -201,13 +198,15 @@ def get_feedback(file: models.File, linter: bool=False) -> _FeedbackMapping:
         in the form ``{line: comment}``.
     """
     res: _FeedbackMapping = {}
+    comments: t.Union[t.List[models.Comment], t.List[models.LinterComment]]
+
     try:
         auth.ensure_can_see_grade(file.work)
 
         if linter:
-            comments = db.session.query(models.LinterComment).filter_by(
-                file_id=file.id
-            ).all()
+            comments = db.session.query(
+                models.LinterComment,
+            ).filter_by(file_id=file.id).all()
 
             for linter_comment in comments:  # type: models.LinterComment
                 line = str(linter_comment.line)
@@ -216,9 +215,9 @@ def get_feedback(file: models.File, linter: bool=False) -> _FeedbackMapping:
                 name = linter_comment.linter.tester.name
                 res[line].append((name, linter_comment))  # type: ignore
         else:
-            comments = db.session.query(models.Comment).filter_by(
-                file_id=file.id
-            ).all()
+            comments = db.session.query(
+                models.Comment,
+            ).filter_by(file_id=file.id).all()
 
             for human_comment in comments:  # type: models.Comment
                 res[str(human_comment.line)] = human_comment
@@ -235,12 +234,19 @@ def delete_code(file_id: int) -> EmptyResponse:
     .. :quickref: Code; Delete the given file.
 
     If a student does this request before the deadline, the file will be
-    completely deleted. , if the request is done after the deadline the user
+    completely deleted. If the request is done after the deadline the user
     doing the delete will be removed from the ownership of the file and if
     there are no owners left the file is deleted.
 
+    If the file owner of the given file is the same as that of the user doing
+    the request (so the file will be completely deleted) the given file should
+    not have any comments (Linter or normal) associated with it. If it still
+    has comments the request will fail with error code 400.
+
     :returns: Nothing.
 
+    :raises APIException: If the request will result in wrong
+        state. (INVALID_STATE)
     :raises APIException: If there is not file with the given id.
         (OBJECT_ID_NOT_FOUND)
     :raises APIException: If you do not have permission to delete the given
@@ -272,6 +278,18 @@ def delete_code(file_id: int) -> EmptyResponse:
     if code.fileowner == other:
         _raise_invalid()
     elif code.fileowner == current:
+        if db.session.query(
+            sql.or_(
+                models.Comment.query.filter_by(file_id=code.id).exists(),
+                models.LinterComment.query.filter_by(file_id=code.id).exists(),
+            )
+        ).scalar():
+            raise APIException(
+                'You cannot delete this file as it has comments',
+                f'The file "{file_id}" has comments associated with it.',
+                APICodes.INVALID_STATE,
+                400,
+            )
         code.delete_from_disk()
         db.session.delete(code)
     elif code.fileowner == models.FileOwner.both:
